removing fatal error
